remove store mocks from hosts tests
