+++ hosts/utils.go
@@ -15,11 +15,11 @@ import (
 	"go.sia.tech/coreutils/rhp/v4/siamux"
 )
 
-type scanner struct{}
+type HostScanner struct{}
 
 // ScanSiamux executes the RPCSettings RPC on the host and returns its settings
 // using the SiaMux protocol.
-func (c *scanner) ScanSiamux(ctx context.Context, hk types.PublicKey, addr string) (proto4.HostSettings, error) {
+func (c *HostScanner) ScanSiamux(ctx context.Context, hk types.PublicKey, addr string) (proto4.HostSettings, error) {
 	t, err := siamux.Dial(ctx, addr, hk)
 	if err != nil {
 		return proto4.HostSettings{}, fmt.Errorf("failed to upgrade connection: %w", err)
@@ -31,7 +31,7 @@ func (c *scanner) ScanSiamux(ctx context.Context, hk types.PublicKey, addr strin
 
 // ScanQuic executes the RPCSettings RPC on the host and returns its settings
 // using the QUIC protocol.
-func (c *scanner) ScanQuic(ctx context.Context, hk types.PublicKey, addr string) (proto4.HostSettings, error) {
+func (c *HostScanner) ScanQuic(ctx context.Context, hk types.PublicKey, addr string) (proto4.HostSettings, error) {
 	t, err := quic.Dial(ctx, addr, hk)
 	if err != nil {
 		return proto4.HostSettings{}, fmt.Errorf("failed to upgrade connection: %w", err)
